PG200309A:
Fixed a minor (but occasionally annoying) problem when showing CTags, in new windows. The cursor was stuck at the beginning (instead of moving to the target line) until some kind of event was triggered in the window, e.g., a mouse click. Now the window is unmapped and remapped which does the trick (forcing an update causing the cursor to move as needed) and additionally raises the window with some extra commotion, so it is brought to the user's attention.

Modified elvis.clr and elvis.ini in data/ to allow spaces in filenames which is the right way to go when a Unix-like system (Cygwin or Linux on a VM) coexists with Windows. Changed the default fonts and theme to what I am used to, so no .elvisrc file is needed to set the preferred defaults.